AllWatcher is now an override rather than a subclass.
Subclasses a facade no longer works.

Of course, now I realize the overrides are broken. That's the next thing
to fix :-)
