--- juju/client/watcher.py
@@ -1,14 +1,0 @@
-from .client import AllWatcherFacade as BaseAllWatcher
-from .client import ClientFacade
-
-
-class AllWatcher(BaseAllWatcher):
-    async def rpc(self, msg):
-        if not hasattr(self, 'Id'):
-            client = ClientFacade.from_connection(self.connection)
-
-            result = await client.WatchAll()
-            self.Id = result.watcher_id
-
-        msg['Id'] = self.Id
-        return await super(AllWatcher, self).rpc(msg)
